requirements.txt
aioboto3==11.2.0
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
astroid==2.15.5
async-timeout==4.0.2
attrs==23.1.0
boto3==1.26.76
botocore==1.29.76
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.8
deeplake==3.6.5
dill==0.3.6
entrypoints==0.4
frozenlist==1.3.3
greenlet==2.0.2
humbug==0.3.1
idna==3.4
isort==5.12.0
jmespath==1.0.1
langchain==0.0.208
langchainplus-sdk==0.0.16
lazy-object-proxy==1.9.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
mccabe==0.7.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numcodecs==0.11.0
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pathos==0.3.0
Pillow==9.5.0
platformdirs==3.7.0
pox==0.3.2
ppft==1.7.6.6
pydantic==1.10.9
PyJWT==2.7.0
pylint==2.17.4
python-dateutil==2.8.2
python-decouple==3.8
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.16
tenacity==8.2.2
tiktoken==0.4.0
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
urllib3==1.26.16
wrapt==1.15.0
yarl==1.9.2
